backend/requirements.txt
six==1.15.0
Django>=3.1.5,<3.1.6
djangorestframework>=3.12.2,<3.12.3
psycopg2>=2.8.6,<2.8.7
Pillow>=8.1.0,<8.2.0
flake8>=3.8.4,<3.9.0
django-cors-headers==3.6.0
django-rest-auth>=0.9.5
django-allauth>=0.41.0,<0.42.0